merging after rollout modus_operandi_release_28 into modus_operandi_hotfix

+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/pickingslotsClearing/process/WEBUI_PickingSlotsClearingView_TakeOutTUAndAddToLU.java
@@ -5,7 +5,7 @@ import de.metas.handlingunits.IHandlingUnitsBL;
 import de.metas.handlingunits.QtyTU;
 import de.metas.handlingunits.allocation.transfer.HUTransformService;
 import de.metas.handlingunits.model.I_M_HU;
-import de.metas.handlingunits.picking.IHUPickingSlotBL;
+import de.metas.handlingunits.picking.slot.IHUPickingSlotBL;
 import de.metas.handlingunits.picking.PickingCandidateService;
 import de.metas.handlingunits.storage.EmptyHUListener;
 import de.metas.process.IProcessDefaultParameter;
@@ -139,7 +139,7 @@ public class WEBUI_PickingSlotsClearingView_TakeOutTUAndAddToLU
 		HUTransformService.builder()
 				.emptyHUListener(EmptyHUListener.collectDestroyedHUIdsTo(huIdsDestroyedCollector))
 				.build()
-				.tuToExistingLU(tuOrAggregatedHU, qtyTU.toBigDecimal(), luHU);
+				.tuToExistingLU(tuOrAggregatedHU, qtyTU, luHU);
 
 		// Remove from picking slots all destroyed HUs
 		pickingCandidateService.inactivateForHUIds(huIdsDestroyedCollector);
